requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
PyYAML==6.0.2
python-dotenv==1.1.1
azure-search-documents==11.5.2
PyJWT[crypto]==2.10.1
requests==2.32.4
msal==1.31.1
python-docx==1.2.0
PyMuPDF==1.26.4
pytest==8.4.1
